Add handicap pairs example and implement support.
The main issue with handicaps is not with them but with adjusted scores.
I don't believe adjustments are reported explicitly by scorer (although
they might be in separate elements -- I don't have any examples as yet).

Hence, in order to report the adjustment value, I compare the raw and
final scores. Obviously this also needs to take the handicap into
account. However we can't just do the calculations and check the
expected and actual final scores are identical as it turns out many are
off by 0.01.

Presumably this is caused by rounding issues. All the values reported
are to two decimal places, but presumably internally scorer is using
more.

For now I'll just treat differences of 0.01 as being the same and
differences of greater magnitude as indicating an adjusted score.