Tricks should be total tricks taken, not the result.
We have to convert from the contract (e.g. "1 NT" or "PASS") and result
(e.g. "+1" or "=") to total number of tricks declarer won.

Aside from "PASS" we also need to handle the case where there is no
contract specified (which can happen if a board is adjusted). I have no
examples with a phantom as yet but presumbably also there.