Divide scorer MPs by 10.
It looks like scorer records MPs as floor(MPs * 10). This causes the
sum of MPs for any given pair to overflow the maximum value in Pianola.
Fix this by simply dividing by 10.

Of course the issue with the values being truncated instead of rounded
remain, but that is much more awkward to deal with, and hopefully won't
matter, so ignore it.