Fixup adjusted traveller line scores.
Scorer is not reporting assigned scores (e.g. A5050) but rather just
"Adj" for the score and then the assigned MPs. This causes problems when
importing into Pianola.

After reading the traveller data we then need to go back and correct any
"Adj" score to something that will be recognised. To determine the
correct value (A5050, A6040, etc) we need to know not only the MPs
awarded but also the total MPs given for the board normally.

In order to calculate this we need to know how many times the board has
been played and the maximum number of times any board in the session has
been played.

This is all a bit involved so for now just use a hard-coded value of
A5050 to get things working.