Calculate adjusted board scores from awarded MPs.
To calculate the awarded score from the awarded MPs we need to know the
total MPs available for the board.

The best way to do this would be to calculate them, but this has proved
difficult to do reliably.

Instead I am using a heuristic and finding all awarded MPs for the
board, counting their occurrence, and using the most common.

This will fail if boards are awarded different MPs in different
sections. That doesn't seem to be how scorer is behaving, at least with
the examples I have, so hopefully it will be reliable.

Additionally this work pointed up a couple of other things that needed
fixing:

1) MPs awarded for pairs were being rounded to an integer. Fix this.
2) MPs awarded for the phantom were hard-coded at -9999 (or -999.9 after
   adjusting for scorer's * 10). Change this to 0.